Finished testing collapsible map thing.
Might need to configure MAX_MAP_OBJECTS.  My local server truncates the page if I do more than 100.
I'm not sure why this is, because it can get away with 1000+ on the production server.  Bear in mind,
though, that we have to use webrick locally, while we use passenger in production.