Big changes to how translations are done:
1.  YAML files are still there, but only the "official" one should be edited by
hand (en-US in our case). 

2.  All the strings are now databased in "languages" and "translation_strings"
tables, but note that Globalite still uses the YAML files. 

3. rake lang:xxx was completely rewritten:

  a) Run "rake lang:update" to update the database and other language files.
  (This replaces the old messy "rake lang:clean".) 

  b) REMEMBER TO RUN "rake lang:update" ON SERVER!!!, after "svn update", if
  you've changed any of the English translations. 

  c) Soon, users will have an interface for editing translations directly on
  the site. 

  d) But users also have the option of editing a plain-text "export" file just
  like before: (Note: lang:update automatically updates all the export files.) 

    rake lang:export:es-ES
    [edit lang/ui/es-ES.txt]
    rake lang:import:es-ES

  e) All the tasks will notify you of any changes to the database, so if you
  don't see anything, it means no strings were changed in the database. 

  f) We should unit test the rake tasks, but I'm not sure how yet.  For now you
  can test it by hand by exporting everything, then importing them and making
  sure it doesn't try to change anything.  Likewise, running lang:update should
  change nothing if no changes were made. 

4.  PLEASE RUN MIGRATION!!  This will populate the languages and
translation_strings tables in your local database. 

5. Note that "translation_strings" is versioned like name_descriptions, etc.

6.  It now tracks the translation strings used on a given page if you are
logged in and reading the site in a language other than English.  It writes a
list of tags used to a temp file, and gives you a link to a page where you can
edit them at the bottom of the page.  (Needs to be written.)  These temp files
are automatically cleaned up every 10 minutes. 

Also in this evision: I split up the api_controller into api_controller and
ajax_controller, refactoring everything in the process.  (Note: we need a lot
more tests of ajax stuff, I'm not 100% sure everything is in top form yet.) 

All tests are passing, but please don't push this out on the production server
until I know I'll be around and can test everything correctly.