Changed how mapping large numbers of objects works.  It now does the queries for observations/locations using select_rows (low level) query, and creates "minimal" observation/location instances.  This allows us to map *all* the results for large queries, such as "map all observations on MO".
Also, it now defers creation of new queries for the subsets of locations under all
the markers.  Before it created the queries early, resulting in the creation of
hundreds of queries for large maps, all of which were hardly ever used.  This was
causing a serious performance hit, too.  Instead it passes the bounding box to
map/show_observations/locations via north/south/east/west parameters, and it
creates a new restricted query only as requested.  Unfortunately this does result
in some confusing behavior when you click on "map" for a set of 19 observations
under a marker, and wind up with a map of 43 observations (because other markers
may be contained within the bounds of the one claiming to have only 19 observations)

Also, I don't know how to deal with observations with bogus lat/long when restricting
like this.  The problem is either we must use a left outer join (ack!) from observations
to locations, or we need to do the query twice, one on observations with a defined
location, one on observations without a defined location.  Then in the one *with*,
we need to do a very complex query telling it to bound the *observation* lat/long
wherever it is at least close to location boundaries, otherwise use the location
boundaries.  I'll look into this nastiness in more detail later maybe.  For now,
it ignores the observation lat/long completely when deciding which observations
occur within a bounding box.  Clearly suboptimal.

It continues to fail the two unit tests for helper/map_colllapsible.rb.