Changes to get image and database backups working
script/update_backup is a minor rewrite of a script I've been using for
years to maintain a backup of the images and database for MO.  It
includes SQL code for striping email addresses out of the database and
will copy over any images on the MO image server (currently at
DreamHost) that are not on the local system.  It does depend on
appropriate ssh keys to exist on both the MO database server and the MO
image server.

I also updated the LICENSE file to 2013.