assert_query tests for subset
Realized that a few tests were failing.  One was just due to a unsafe
print statement I'd added that I just removed.  The other was due to
changes in the fixtures impacting the query results.  Fixed that by
having assert_query check that the expected results are a subset rather
than equal to the results.  This caused no new failures.