Make assert_response_equal_file work again (Also require cvs in ApplicationController rather than individual controllers.)
Assertion modified because ActionDispatch::Response#body behaves differently in Rails 4 than Rails 3.
In Rails 3, if I used cvs to generate a string, and then sent it to a file, response.body would return the raw string.
In Rails 4, response.body strips `\n` used to separate rows.
So in order to get the `\n`s back, I join the body_parts with `\n`.
Ugly, but nothing better occurs to me.