Second draft of automatic sitemap generator.
It now additionally creates robots.txt and a parallel "dumb" index to ensure
that robots can reach all the pages in the sitemap.  This robot index is
reached when a robot follows the link under the "Mushroom Observer" title in
the top left corner of any page.  Non-robots will continue to be taken to the
list_rss_logs "home page" like normal.