Change nokogiri version (and dependencies).
I am doing this because Travis is erroring with this message:

Your bundle is locked to nokogiri (1.6.8.1), but that version could not be found
in any of the sources listed in your Gemfile. If you haven't changed sources,
that means the author of nokogiri (1.6.8.1) has removed it. You'll need to
update your bundle to a different version of nokogiri (1.6.8.1) that hasn't been
removed in order to install.