Second draft.  Should be mostly working.
I ripped out the name, user and species_list parameters on one of the observation queries.  It doesn't seem to have been used by anything.  And if we actually want this, then we should move it to observation_filters to give all observation queries access to them.

I totally redid how sort order and title are dealt with in coercion.  It is much simplified and behaves very similarly in practice to the old code.

All tests in test/model/query_test.rb are passing.  It's a good start...