Include “group” at end of real_text_name if display_name ends in “group”
Name#display_to_real_text was stripping “group”  from real_text_name if “group” was at the end of display_name.

But “group” **must** be included in the real_text_name of a group because: (a) real_text_name is what appears in the edit_name form's Scientific Name field, and (b) “group” must appear there in order to keep the name consistent.