Revise #real_text_name
`real_text_name` is derived on the fly from `display_name`.
For “group” names, `display_name` now includes “group” and so should `real_text_name`.
But the first `sub` in `real_text_name` removes everything after the taxon, including “group”.
So the revision is to add “group” if it was removed.