Remove first edit_name call to find_or_create_name_and_parents
This is the call which caused problems by, among other things, preventing an edit from removing a Name’s author.
It’s replaced by `another_existing_name` which simply returns the “new name”, without checking for authors, finding ancestors, or saving anything.

The points are:
(a) for at least Names at the :Group level, we want authored and unauthored Names to coexist, and
(b)  when editing (as opposed to creating) a Name, the user first selects a Name, and we presume the user wants to edit that Name alone.