Deny write permission if article Project doesn’t exist
The Project which controls Article write permission does not exist in the development or production databases; it’s not created by this branch, so it must be created in those live db’s. Until then, the app should not throw errors when trying to figure out if a user has write permission.  (We also don’t want errors thrown if the Project gets deleted.) Therefore, if the project doesn’t exist, just deny permission.