Make notes default NULL in Observation fixture
After the notes format is changed, and notes are serialized, both nil and "" notes will be persisted as NULL.
Therefore current "" and nil notes will be up-migrated to NULL.
Because there will then be no way to tell them apart, they will both be NULL after down-migration. So then there's no way to tell them apart.
So we might as well make the default NULL to start with.