Fix sass 0px == 0 deprecation warning
When starting the server, there are a bunch of warnings like
>DEPRECATION WARNING on line 339 of /vagrant/mushroom-observer/app/assets/stylesheets/mushroom_observer.scss:
The result of `0px == 0` will be `false` in future releases of Sass.
Unitless numbers will no longer be equal to the same numbers with units.

The fix is to compare units to units, i.e., 0px == 0px

Delivers https://www.pivotaltracker.com/story/show/115775707 [Delivers #115775707]