Only nucleotides in BLAST's QUERY parameter
If I simply wrap the Bases in quotes and use that as NCBI's QUERY param, then NCBI can mess up some of the statistics.
(NCBI will BLAST the QUERY, and correctly report the identity statistic (Ident). But can mess up other statistics.)
So the best approach is to use only the nucleotide codes as the QUERY param, removing digits and whitespace.

- Add and modify tests and fixtures
- Modify Sequence so that tests pass
- Also remove whitespace when Accession is used as QUERY param
- Add Constant and adjust constant names to be more precise
- change name of `blastable` validator method to avoid confusion and potential collision with `blastable?` method