Fix default app_index_a_z translation
Make it something which (sort of) actually appears in the lh navbar
Old default (Index A-Z) hasn't been in the navbar for a long time.
So pages (e.g. Spanish translation) which referenced this label had useless text.
The current equivalent is the Names link under Index. So make "Index Names" the default.