Ran bundle update, but held back the mail gem
mail 2.7.1 adds =0D to the end of lines in our text emails.
It's not clear if this will cause problems for the users who
are currently receiving text emails or not.  2.7.0 does not
appear to suffer from this problem, so for now we are locking
on that version of the gem.