[Delivers #163597896] Fix Error on adding Comment to Name
Adding a Comment to a Name that has synonyms was throwing an error.
The Comment form tries to create a link to all the synonyms.
It uses a partial to do this: app/views/name/_name_html.erb
Rails 4 `link_to` used to create the link route with the :name controller.
But Rails 5 does it with the :comment controller in this situation.
Solution:
- specify `controller: :name` in calls to `link_with_query` in _name_html.
- I also did some formatting changes in a couple of views so that I could read them easier. (Shortening the lines, and aligning the Rails stuff.)