[Fixes #167543713]Fix timing issue in test_blocked_ips
Address [Pivotal #167543713](https://www.pivotaltracker.com/story/show/167543713)

RobotsTest#test_blocked_ips was failing intermittently. The cause seems to be  :
(a) modifying a file which lists  ok ips
too soon after
(b) the modification in memory of a list of blocked ips.
I made  the test pass by adding a 1 second delay between modification of the list in memory and the file.