Remove former name of cop from disable/enable
Tests if syncing Codeclimate channel to RuboCop version solves a Codeclimate problem

The correct name of this cop is `HashAlignment`. But using that name created Codeclimate issues.
Therefore @mo-nathan used both the current and former (`AlignHash`).

One suspect cause is: Codeclimate did not use the RuboCop version specified in Gemfile.
The Codeclimate channel and RuboCop version are now synced.
So this commit tests if this suspicion is correct.