Turn UsersControllerTest#test_index_user green
Fixes a bug introduced by the transition from `master` to `bootstrap-4`.

For a non-admin user, `index` redirects to `:rss_logs`
In master, `:root` was `:rss_logs`, so the test assertion worked.
But in the current version of `bootstrap-4`, `:root` is `observations#index`. [Why the swtich?]
So I need to change either `:index` or the assertion. I picked the latter in order to maintain
the same behavior for `:index`.