Turn UsersControllerTest#test_show_no_id green
Changed assertions to match changed behavior in bootstrap4
In `master`, `get(:show_user)` without an id  was (somehow) redirected to `:index_user`.
(At the moment I can't tell how because I've lost my connection.)
In `bootstrap-4`, there's no route to `get(:user)` without id. So that request:
- Throws an error.
- Returns `:success` with an empty response.body.
I think that's bad. (Or maybe it's a better way to handle robots.)
Could we restore the old behavior by adding route so that `get(:user)` redirects to users/index_user?
But if the requesting user is not an admin, that will redirect to :rss_logs.