Fix custom_file_field, updated for bootstrap-4 with new js
The new glossary term form needed the application helper method "custom_file_field", which was broken.

This commit repairs that method for bootstrap-4 and adds the necessary js dependency and initializer call.

In an upcoming PR this method should be deployed in the Image, Account and Observation forms too. (Possibly other forms as needed, TBD).