Have it supply default behavior for include_synonyms and include_subtaxa separately. The use case that convinced me is:
Search for "Something", but this isn't a name so it treats it as a pattern search.
You realize that you didn't want this to include subtaxa, so you include "include_subtaxa:no".
Now it interprets this as a name search because of the subtaxa modifier, but now it also includes synonyms be default.
It seems weird that by adding "include_subtaxa:no" it would also unexpectedly add "include_synonyms:yes".

This is probably a bad example, because of course, searching for "Something" shouldn't actually result in a well-defeined name search, so why would the user decide that they didn't want subtaxa of "Something"?  Presumably because there were strange results because it wasn't actually searching for Something the name!
But in general, this just seems to make more sense to me -- that it should provide defaults for all of the taxa modifiers:

include_synonyms -> yes
include_subtaxa -> yes
include_all_name_proposals -> no
exclude_consensus -> no

In practice, only the "yes" defaults actually need to be considered.  But I think it makes sense, when thinking about it like this, to supply defaults individually.