Update rss_log_test.rb
This added test is a bit much, but i'm honestly not sure what an adequate level of coverage is! If someone has an opinion, please suggest shortening it. The second thru 4th assertions seem like the core of it. I don't understand why the notes for the "albion" log are "Updated Created", but that's from the fixture.