Radical simplification of User.primer_data from Arel to AR/ruby
Moved the concat of login + <name> out of the SQL, so it's just returning the two columns from the records. Ruby handles the conditional concat. The Arel conditional orderby :last_login is also shortened with much more legible Arel extensions. CASE statement replaces IF because that's what the extensions provide.