Fix form selects (with class)
To reproduce: go to your account preferences, select the Amanita theme, and reload the page. Preference selects are not really very legible; insufficient contrast.

Diagnosis:
In the view templates, form selects (in user prefs and elsewhere) correctly specify a Bootstrap `class: "form-control" but the class is not applied due to incorrect syntax in the erb tag. Result: the select options are not very legible; the text color is incorrectly inherited from the parent element.

Fix: The `form.select` helper requires two options hashes. The `class` has to be in the second hash, unintuitively. Someone seems to have discovered this already, because form_name already had the correct syntax.