Good Lord. CRUDify observation map
This one commit necessitated a lot of salutary changes to the code. The main one restores Joe's discovery that **in the context of namespaced controllers**, `show_controller` always needs a leading forward slash to specify a top-level controller. I had foolishly reverted this awhile back and everything was working ok, but that's because i hadn't namespaced enough controllers yet.