Make community votes table a popup with live data
Before, the community vote table was pre-rendered in the page with the data on page load. But voting via Ajax would not affect the table; it would always display the votes that were originally there on page load.

Now, the vote table is rendered by `VotesController#show`, as it would be for no-js users, and therefore should show live data at the time of the request.