Flip terminology: `tab` = link attribute array, `link` = HTML link
MO makes pretty good use of a type of object i’m calling “link attribute arrays” in the helpers, things like this:
   [label, path, **args]
(I can’t remember if MO used the args much, but i started adding them so i could put identifying classes on links for integration tests). These objects are not quite links, but they have all the attributes one needs to pass to `link_to` or `button_to` to create HTML for the templates.

Confusion arises, however, when passing these objects around in helpers. Some objects called "blahblah_link" being passed around are in fact links, while most are simply these attribute arrays. This PR hopes to make it more obvious which is which at a glance. If only we can get used to the terminology.

Up til now, the term `tab` was used for links in the title bar, making up a "tab set". I'm going to move to call these "link sets" and try to reserve the word `tab` for methods that pass these link attribute arrays around.

In the wild, `tab` generally refers to a JS navigation pattern where different panels of info in the same area of the page are shown/hidden by clicking labels at the top or side of the panels. Neither of our "tabs" are these type of "tab"