Set modal form titles
The recent refactor eliminated the `@title` ivar, which was set in the controller for these actions. The modal forms were depending on that ivar for their own titles, since they don't have independent views to set it: they use a shared modal form rendering partial.

This PR moves the title setters to helpers, and then sends them via the controller to the shared partial. Is that more convoluted than an ivar? Yes. I didn't realize this when i set about refactoring titles.

The better solution is via Trailblazer, but we're not there yet.