Host jquery locally
This is to shut up the people who have installed browser extensions blocking jquery and then tell us that the site doesn't work, mysteriously.

I'm reluctant to do this because it locks our version of a frequently changing library, and forces us to manually update the library. At a certain point we're going to have to tell people to stop blocking public JS repos, or sorry Charlie.

But maybe this is not so bad in this case, because we're aiming to eliminate the jQuery dependency with Bootstrap 5, so maybe it's just for a few months.