Move `read_exif_data` to Image as an instance method
Creates a method to read the image's exif data, from what had been in the show action of the Images::EXIFController. This is an AJAX response to show an image's exif header. The method reads the info off the actual file on the server; we do not store this info in the Image record.

PR makes the method available outside that controller, in other words.

Also clears up a method name that I hadn't understood:
- rename Image::URL method from `file_name` to `full_filepath`
- rename Image instance method `local_file_name` to `full_filepath`

These methods return the full path of the image on the server per environment, like this:
`/Users/nimmo/Documents/GitHub/developer-startup/mushroom-observer/public/images/orig/1663403.jpg`