Correct the field name for `accession_number` on the obs form
We've been calling this field `herbarium_id` as a shorthand, but it's actually the field for the HR `accession_number`. This PR just renames the field and the ivar `@herbarium_id` generated in the context of the `ObservationsController` (only).

`herbarium_id` would conflict with the new hidden field for autocompleters that is generated with the more accurate use of the name `herbarium_id`.