Fix "Turn off admin mode" link in nav menu
Fixes #2168. Adds test. @JoeCohen - In case you tried to fix this and got stumped, it wasn't working because this link needs to submit a "POST" request.

Here's how to build links like this:
- the link must use `button_to`, rather than a `link_to`
- the button needs a couple extra classes to style it like an `<a>` if that's what we want

Rails supposedly allows `link_to` with `data: { turbo_method: :post }` but in reality (at least on MO) it doesn't work, the CSRF stops the link from working.

Also, the [Turbo docs themselves state](https://turbo.hotwired.dev/handbook/drive#performing-visits-with-a-different-method) "you should also consider that for accessibility reasons, it’s better to use actual forms and buttons for anything that’s not a GET." The Rails `button_to` helper builds a form and a button, as you probably know.