Fix the date select helper
The helper I wrote wasn't handling options right, which is no doubt why @mo-nathan or @jdcohenesq couldn't use it on the Field Slip form.

This fixes the option handling and updates the form to use our formatted date select.

(The issue was that the field name `date` used in the form is not an attribute of a `FieldSlip`. This fixes it so the helper tries to access the attribute to get a date, but falls back to today's date, which is what the current form field defaults to.)