Rename Query param[:by] as [:order_by]
On the AR branch, I'm thinking to move the ordering logic into scopes, with the current param[:by] delegating to a new AbstractModel scope `.by` as the dispatcher to more specific ordering scopes in each appropriate model.

However, the name `:by` seems a little ambiguous in this case. The meaning of `:order_by` is clearer.