Rename pagination stuff
The ubiquitous ivar @pages has confused me for years. Until today thought it contained query results, and I can't remember the last time I really checked what it contains.

It's just a little package of data for the pagination nav. It gets populated by ApplicationController using the query, but that's different.

I also thought `paginate_block` et al. did a bunch of magical stuff with the query, but it just adds a pagination nav on the top and bottom of the results. Yes, that's a valid meaning of "paginate", but in the context of a query "paginate" means "return results subdivided by a given page".

This PR finally renames what has confused me for years, hopefully it demystifies these methods for others.