Fix query_param called by Observation index param `where`
`search_where` is the correct Query param and scope name for a text fuzzy search of observations by location name (which gets undefined locations). The scope cannot be called `where`, because that refers to the AR method.

The scope called by this index param was incorrectly set by me to `locations`, which only searches the location table.