Reset Stimulus "connected" convention
Currently, if two stimulus controllers are applied on a single element, they will complete to apply the same "data-stimulus" attribute. This is not helpful, as Nathan and I noticed when writing his "search-status_controller".

This PR follows the new example we made there, and gives each stimulus controller its own data attribute name so we can tell if the element is connected. e.g. `data-field-slip-job="connected"` instead of `data-stimulus="field-slip-job-connected"`.