Remove all double-spaces from translation strings
`Double spaces after a period.  (Like this.)  Or this.` despite being good practice in typing class, are apparently not rendered on the front-end by Textile, they seem to get "squeezed" to a single space.

The presence of double spaces in the original multi-sentence translation string makes testing for the presence of the string impossible without string manipulation or regex, and this seems silly. The string should at least match itself as rendered.

This PR proposes to get rid of the issue at the source, so we can write simpler tests.