Add "Actions" to the mobile menu
On mobile, back when "Actions" was rendered as a plain list of links in the "tabset", it simply took up a big chunk of the screen at the top of every page. On a phone, this could cover most of the screen.

Moving it to a dropdown menu from the top_bar made it inaccessible on mobile. This PR adds the "Actions" menu items to the top of the sidebar on mobile.

More work could be done, but this at least restores functionality.