Fix new Phlex email form params in controllers
Most of our forms concern db models. But some of our older forms just send params.

Phlex superforms all need to use a form object - whether it's the database record or a PORO (using ActiveModel), superform wants to validate the object you're creating/updating when you send the form, as it should - this is a sturdier code pattern than validating params individually. The params become attributes of that form object.

So the Phlex conversion has involved creating objects for object-less forms. But claude and i forgot to update the controllers to look for the params nested under these objects. This PR fixes that.

This "form-object" thing is incidentally the same pattern we now use on Search Forms.